smitebot: move setup_nyx into utils
bench-exec needs the same Nyx sharedir setup as start, so lift the helper
out of start.rs into utils instead of duplicating it.
